refactor(DatasetExplorer): extract hooks and utilities, reduce to 624 lines (#123)
Phase 5 of the DatasetExplorer refactoring. Extract remaining logic into
3 new hooks (useFilterState, useCountBy, useDatasetLoader) and a utils
module. Delete 97-line dead renderCountIndicator function. Add 66 new
tests. Update CORE.md with architecture documentation.

- DatasetExplorer.tsx: 1,926 → 624 lines (-68%)
- Cumulative: 6,175 → 624 lines (-90% across Phases 1-5)
- Tests: 163 → 229 client tests (66 new)
